:truck: Move bucket utils to bucket app
